Fix some ubifs bugs. 1. Don't use spinlock when printing ubifs debug message, or it will cause spinlock recruision; 2. Merge a ubifs patch from kernel 3.10: UBIFS: fix a horrid bug
Al Viro pointed me to the fact that '->readdir()' and '->llseek()' have
no
mutual exclusion, which means the 'ubifs_dir_llseek()' can be run while
we are
in the middle of 'ubifs_readdir()'.

This means that 'file->private_data' can be freed while
'ubifs_readdir()' uses
it, and this is a very bad bug: not only 'ubifs_readdir()' can return
garbage,
but this may corrupt memory and lead to all kinds of problems like
crashes an
security holes.

This patch fixes the problem by using the 'file->f_version' field, which
'->llseek()' always unconditionally sets to zero. We set it to 1 in
'ubifs_readdir()' and whenever we detect that it became 0, we know there
was a
seek and it is time to clear the state saved in 'file->private_data'.

I tested this patch by writing a user-space program which runds readdir
and
seek in parallell. I could easily crash the kernel without these
patches, but
could not crash it with these patches.

Change-Id: I7ec4a9c0240fe377e6a0408e67a17620443d0681